pyproject.toml
[tool.poetry]
name = "footlocker-verify"
version = "0.1.0"
description = "Automatically logs into footlocker owned sites, pulls info from a csv file, and updates user/payment info."
authors = ["Wes Ruedlinger <wes.ruedlinger91@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
argparse = "^1.4.0"
requests = "^2.27.1"
pandas = "^1.4.2"
asyncio = "^3.4.3"
tqdm = "^4.64.0"
playwright = "^1.22.0"
python-decouple = "^3.6"
discordwebhook = "^1.0.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pre-commit = "^2.19.0"
mypy = "^0.960"
flake8 = "^4.0.1"
shiv = "^1.0.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
